Search feature works, filter not yet
