+++ src/views/ViewContext.java
@@ -15,5 +15,7 @@ public enum ViewContext {
 	TASK_LIST_ON_DATE,
 	
 	/** The task is being displayed from the calendar view (e.g. as a popup). */
-	CALENDAR
+	CALENDAR,
+	
+	SEARCH
 }
